LayoutTests/editing/selection/modal-dialog-select-paragraph-expected.txt

This text should be selectable with the mouse.

PASS: all dialog text is selected
